int conversion

+++ codeartproject/projectFiles/admin_credentials.py
@@ -5,7 +5,7 @@ import uuid
 
 string_uuid = str( uuid.uuid4() )
 print(string_uuid)
-half_len_of_string = len(string_uuid)/2
+half_len_of_string = int( len(string_uuid)/2 )
 print(half_len_of_string)
 half_string = string_uuid[:half_len_of_string]
 created_admin_id = int( half_string )
